backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
qrcode==7.4.2
razorpay==1.4.1
python-decouple==3.8
drf-yasg==1.21.7
